# apps/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend API for QuokkaQ, rewritten in Go.",
        "title": "QuokkaQ Go Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3001",
    "basePath": "/",
    "paths": {
...
        "/units": {
            "get": {
                "description": "Retrieves a list of all units",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "units"
                ],
                "summary": "Get all units",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Unit"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new unit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "units"
                ],
                "summary": "Create a new unit",
                "parameters": [
                    {
                        "description": "Unit Data",
                        "name": "unit",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Unit"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Unit"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/units/{id}": {
            "get": {
                "description": "Retrieves a specific unit by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "units"
                ],
                "summary": "Get a unit by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unit ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Unit"
                        }
                    },
                    "404": {
                        "description": "Unit not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a unit by its ID",
                "tags": [
                    "units"
                ],
                "summary": "Delete a unit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unit ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates an existing unit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "units"
                ],
                "summary": "Update a unit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unit ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Unit Data",
                        "name": "unit",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Unit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Unit"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/units/{unitId}/ad-settings": {
            "patch": {
                "description": "Updates ad settings for a unit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "units"
                ],
                "summary": "Update ad settings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unit ID",
                        "name": "unitId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Ad Settings",
                        "name": "settings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/units/{unitId}/materials": {
            "get": {
                "description": "Retrieves materials for a unit",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "units"
                ],
                "summary": "Get unit materials",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unit ID",
                        "name": "unitId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UnitMaterial"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Adds a material to a unit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "units"
                ],
                "summary": "Add material to unit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unit ID",
                        "name": "unitId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Material Data",
                        "name": "material",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UnitMaterial"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.UnitMaterial"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/units/{unitId}/materials/{materialId}": {
            "delete": {
                "description": "Deletes a material by ID",
                "tags": [
                    "units"
                ],
                "summary": "Delete material",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unit ID",
                        "name": "unitId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Material ID",
                        "name": "materialId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Booking": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "scheduledAt": {
                    "type": "string"
                },
                "serviceId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                },
                "userPhone": {
                    "type": "string"
                }
            }
        },
        "models.Counter": {
            "type": "object",
            "properties": {
                "assignedTo": {
                    "type": "string"
                },
                "assignedUser": {
                    "$ref": "#/definitions/models.User"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                }
            }
        },
...
        "models.PreRegistration": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "6-digit unique code",
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "customerName": {
                    "type": "string"
                },
                "customerPhone": {
                    "type": "string"
                },
                "date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "service": {
                    "$ref": "#/definitions/models.Service"
                },
                "serviceId": {
                    "type": "string"
                },
                "status": {
                    "description": "created, canceled, ticket_issued, completed",
                    "type": "string"
                },
                "ticket": {
                    "$ref": "#/definitions/models.Ticket"
                },
                "ticketId": {
                    "type": "string"
                },
                "time": {
                    "description": "HH:MM",
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.Role": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Service": {
            "type": "object",
            "properties": {
                "backgroundColor": {
                    "type": "string"
                },
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Service"
                    }
                },
                "description": {
                    "type": "string"
                },
                "descriptionEn": {
                    "type": "string"
                },
                "descriptionRu": {
                    "type": "string"
                },
                "duration": {
                    "description": "In seconds",
                    "type": "integer"
                },
                "gridCol": {
                    "type": "integer"
                },
                "gridColSpan": {
                    "type": "integer"
                },
                "gridRow": {
                    "description": "Grid configuration",
                    "type": "integer"
                },
                "gridRowSpan": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "isLeaf": {
                    "type": "boolean"
                },
                "maxWaitingTime": {
                    "description": "In seconds",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "nameEn": {
                    "type": "string"
                },
                "nameRu": {
                    "type": "string"
                },
                "numberSequence": {
                    "type": "string"
                },
                "parentId": {
                    "type": "string"
                },
                "prebook": {
                    "type": "boolean"
                },
                "prefix": {
                    "type": "string"
                },
                "textColor": {
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                }
            }
        },
...
        "models.SlotConfig": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "days": {
                    "description": "[\"monday\", \"tuesday\", ...]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "endTime": {
                    "description": "HH:MM",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "interval": {
                    "description": "in minutes",
                    "type": "integer"
                },
                "startTime": {
                    "description": "HH:MM",
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.Ticket": {
            "type": "object",
            "properties": {
                "booking": {
                    "$ref": "#/definitions/models.Booking"
                },
                "bookingId": {
                    "type": "string"
                },
                "calledAt": {
                    "type": "string"
                },
                "completedAt": {
                    "type": "string"
                },
                "confirmedAt": {
                    "type": "string"
                },
                "counter": {
                    "$ref": "#/definitions/models.Counter"
                },
                "counterId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isEod": {
                    "type": "boolean"
                },
                "lastCalledAt": {
                    "type": "string"
                },
                "maxWaitingTime": {
                    "description": "Snapshot from Service at creation",
                    "type": "integer"
                },
                "preRegistration": {
                    "description": "No DB FK: avoids AutoMigrate cycle with pre_registrations.ticket_id → tickets.id",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.PreRegistration"
                        }
                    ]
                },
                "preRegistrationId": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "queueNumber": {
                    "type": "string"
                },
                "service": {
                    "$ref": "#/definitions/models.Service"
                },
                "serviceId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "ttsUrl": {
                    "description": "URL to the generated TTS audio file",
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                }
            }
        },
        "models.Unit": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "companyId": {
                    "type": "string"
                },
                "config": {
                    "type": "object"
                },
                "counters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Counter"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "preRegistrations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PreRegistration"
                    }
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Service"
                    }
                },
                "slotConfig": {
                    "$ref": "#/definitions/models.SlotConfig"
                },
                "tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Ticket"
                    }
                },
                "timezone": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.UnitMaterial": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "roles": {
                    "description": "Relations",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserRole"
                    }
                },
                "type": {
                    "type": "string"
                },
                "units": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserUnit"
                    }
                }
            }
        },
        "models.UserRole": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/models.Role"
                },
                "roleId": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}